+++ .php-cs-fixer.dist.php
@@ -1,8 +1,6 @@
 <?php
 
-use PhpCsFixer\ {
-    Finder, Config
-};
+use PhpCsFixer\{Config, Finder};
 
 $finder = Finder::create()
        ->in(__DIR__ . '/src')
@@ -19,6 +17,8 @@ return $config
     ->setUsingCache(false)
     ->setRules([
                    '@PSR2'                        => true,
+                   '@PSR12'                       => true,
+                   '@PHP74Migration:risky'        => true,
                    '@Symfony'                     => true,
                    'array_syntax'                 => ['syntax' => 'short'],
                    'binary_operator_spaces'       => [