+++ src/Export/Field/Attribute.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Omikron\FactFinder\Oxid\Export\Field;
 
 use OxidEsales\Eshop\Application\Model\Article;
-use OxidEsales\Eshop\Application\Model\Attribute as ArticleAttribute;
+use OxidEsales\Eshop\Core\DatabaseProvider;
 
 class Attribute implements FieldInterface
 {
@@ -29,9 +29,42 @@ class Attribute implements FieldInterface
 
     protected function getData(Article $article): array
     {
-        $attributes = $article->getAttributes()->getArray();
-        return array_reduce($attributes, function (array $result, ArticleAttribute $attribute) {
-            return $result + [$attribute->oxattribute__oxtitle->rawValue => $attribute->getFieldData('oxvalue')];
+        $attributes = $this->getAttributes($article->getId(), $article->getParentId());
+
+        return array_reduce($attributes, function (array $result, array $attribute) {
+            return $result + [$attribute['OXTITLE'] => $attribute['OXVALUE']];
         }, []);
+    }
+
+    protected function getAttributes($articleId, $parentId): array
+    {
+        $oDb                = DatabaseProvider::getDb(DatabaseProvider::FETCH_MODE_ASSOC);
+        $oxAttribute        = getViewName('oxattribute');
+        $oxobject2Attribute = getViewName('oxobject2attribute');
+
+        $select = "select {$oxAttribute}.`oxid`, {$oxAttribute}.`oxtitle`, o2a.`oxvalue` from {$oxobject2Attribute} as o2a "
+            . "left join {$oxAttribute} on {$oxAttribute}.oxid = o2a.oxattrid "
+            . "where o2a.oxobjectid = :oxobjectid and o2a.oxvalue != '' "
+            . "order by o2a.oxpos, {$oxAttribute}.oxpos";
+
+        $articleAttributes = $oDb->getAll($select, [':oxobjectid' => $articleId]);
+
+        if ($parentId) {
+            $parentAttributes = $oDb->getAll($select, [':oxobjectid' => $parentId]);
+            return $this->mergeAttributes($articleAttributes, $parentAttributes);
+        }
+
+        return $articleAttributes;
+    }
+
+    protected function mergeAttributes($articleAttributes, $parentAttributes): array
+    {
+        if (!count($parentAttributes)) {
+            return $articleAttributes;
+        }
+
+        return array_values(array_reduce($articleAttributes + $parentAttributes, function (array $added, array $attribute) {
+            return in_array($attribute['OXID'], $added) ? $added : $added + [$attribute['OXID'] => $attribute];
+        }, []));
     }
 }